# Optimize concrete name scopeLookup