Old scope lookup merged all scopes (current + parents), and then looked up a name. If we have more than one scope, that's linear in the total size of all scopes, because we take unions of Map-s.

The new implementation looks up the name separately from all scopes without merging. That's O(log n * number of scopes) instead.

Relevant changes in `make stdlib-test` times:

```
Total                                   234,777ms
Scoping                                   3,245ms (15,353ms)
Parsing                                   3,119ms (16,003ms)
Parsing.OperatorsExpr                    10,680ms
Parsing.OperatorsPattern                  2,203ms

```

New:

```
Total                                   228,154ms
Scoping                                   2,728ms (14,535ms)
Parsing                                   3,064ms (12,063ms)
Parsing.OperatorsExpr                     7,439ms
Parsing.OperatorsPattern                  1,558ms
```

Probably a lot of name lookup is counted under operators parsing.